requirements.txt
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
configparser==3.7.1
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
entrypoints==0.3
functools32==3.2.3.post2
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
ipython-genutils==0.2.0
Jinja2==2.10.1
jsonschema==2.6.0
jupyter-core==4.4.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mistune==0.8.4
more-itertools==5.0.0
nbconvert==5.4.0
nbformat==4.4.0
ndg-httpsclient==0.5.1
numpy==1.16.1
pandas==0.24.1
pandocfilters==1.4.2
pathlib2==2.3.3
pickleshare==0.7.5
prompt-toolkit==1.0.15
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
pymongo==3.7.2
pyOpenSSL==19.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.9
Quandl==3.4.8
requests==2.22.0
scandir==1.9.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
sklearn==0.0
SpeechRecognition==3.8.1
sqlparse==0.3.0
subprocess32==3.5.4
testpath==0.4.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
psycopg2-binary==2.8.3
